Rename landing-page components
